splat function

+++ src/SplatMotionsWrapper.cpp
@@ -13,7 +13,7 @@ void SplatMotionsWrapper(double *vx, double *vy,
                          const double *vxForward, const double *vyForward,
                          const double *vxBackward, const double *vyBackward,
                          const double *Im1, const double *Im2,
-                         int colType, int h, int w, int c, float t) {
+                         int colType, int h, int w, int c, double t) {
   DImage ImFormatted1, ImFormatted2;
   DImage vxForwardFormatted, vyForwardFormatted;
   DImage vxBackwardFormatted, vyBackwardFormatted;
